move more implementations from h to cpp
